Create a new session object for each session.
To stop state leaking from one session to the next, do not create the
session object until the user is about to switch from the asleep
state to the login state. This session object will exist until the
use logs out but the next time e login beings, it will be with a new
session object.