Rework handling of pending new lines in PRINT
Determine this at parse time and add a newline item to the end of
the print items if needed. Remove all the "pending" handling in
the terminal channel code.

Rework the handling of commas to correctly start a zone on the next
line if it won't fit on the current one.