Introduce vitrual machine
Partway through implementing function definition blocks, it is clear
that the execution by evaluation of the parse tree breaks down when
a function definition block called from an expression performs input:

100 DEF FNX(A$)
110 PRINT A$
120 INPUT N
130 FNX=N
140 FNEND
...
200 PRINT SQR(FNX("ENTER A POSITIVE NUMBER"))

So we're forces to tick finer grained things than statements. Compile
into a virtual machine and tick those operations. This commit is the
first step, with PRINT, LET and END working for numeric scalars.