Fix bug where GOTO out of FOR left the frame on th stack even when the same loop was started from the top.
Now, every time we start a FOR loop, we look to see if the loop is
already active and if it is we wipe out all the contexts up to and
including that one and start again. This should allow us to recurse
into DEF blocks that contain FOR loops if the loop control variable
is local.